Remove deprecated subscribe
