+++ lib/tackle.rb
@@ -1,44 +1,22 @@
 require "tackle/version"
+require "bunny"
+require "logger"
 
 module Tackle
-  require "tackle/worker"
+  require "tackle/connection"
   require "tackle/publisher"
   require "tackle/consumer"
 
-  def self.consume(params = {}, &block)
+  module_function
+
+  def consume(params = {}, &block)
     params   = Tackle::Consumer::Params.new(params)
     consumer = Tackle::Consumer.new(params)
 
     consumer.subscribe(&block)
   end
 
-  # deprecated
-  def self.subscribe(options = {}, &block)
-    # required
-    exchange_name = options.fetch(:exchange)
-    routing_key   = options.fetch(:routing_key)
-    queue_name    = options.fetch(:queue)
-
-    # optional
-    amqp_url    = options[:url]
-    retry_limit = options[:retry_limit]
-    retry_delay = options[:retry_delay]
-    logger      = options[:logger]
-    on_uncaught_exception = options[:on_uncaught_exception]
-
-    worker = Tackle::Worker.new(exchange_name,
-                                routing_key,
-                                queue_name,
-                                :url => amqp_url,
-                                :retry_limit => retry_limit,
-                                :retry_delay => retry_delay,
-                                :logger => logger,
-                                :on_uncaught_exception => on_uncaught_exception)
-
-    worker.subscribe(&block)
-  end
-
-  def self.publish(message, options = {})
+  def publish(message, options = {})
     # required
     exchange_name = options.fetch(:exchange)
     routing_key   = options.fetch(:routing_key)
