--- lib/tackle/tackle_logger.rb
@@ -1,13 +1,0 @@
-module Tackle
-
-  module TackleLogger
-
-    def tackle_log(message)
-      pid = Process.pid
-
-      whole_message = "tackle - pid=#{pid} message=#{message}"
-
-      @logger.info(whole_message)
-    end
-  end
-end
