+++ lib/tackle/version.rb
@@ -1,3 +1,3 @@
 module Tackle
-  VERSION = "0.9"
+  VERSION = "1.0"
 end
